# ADO.EntityFrameworkCore/StatsModel.cs
namespace ADO.EntityFrameworkCore.Models
{
    public class StatsModel
    {
        public double containers_dispatched { get; set; }
        public double containers_not_dispatched { get; set; }
        public double budget_used   { get; set; }
    }
}