# ContenedoresRest/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ContenedoresRest, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Containers": {
      "post": {
        "tags": [
          "Containers"
        ],
        "operationId": "selectContainers",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "budget",
            "type": "number",
            "format": "double"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ContainersModel"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/Stats": {
      "get": {
        "tags": [
          "Stats"
        ],
        "operationId": "stats",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/StatsModel"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ContainersModel": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "transportCost": {
          "format": "double",
          "type": "number"
        },
        "containerPrice": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
    "StatsModel": {
      "type": "object",
      "properties": {
        "containers_dispatched": {
          "format": "double",
          "type": "number"
        },
        "containers_not_dispatched": {
          "format": "double",
          "type": "number"
        },
        "budget_used": {
          "format": "double",
          "type": "number"
        }
      }
    }
  }
}